BSAS-20 [+] modify 'int id' -> 'Long id'

+++ engine/src/main/java/com/jsoftware/platform/controller/RecommendHotspotController.java
@@ -26,7 +26,7 @@ public class RecommendHotspotController {
     }
 
     @GetMapping("/api/recommendHotspot/{id}")
-    public RecommendHotspot getRecommendHotspot(@PathVariable("id") int id) {
+    public RecommendHotspot getRecommendHotspot(@PathVariable("id") Long id) {
         System.out.println(id);
         return activityService.readRecommendHotspot(id);
     }
